Make header only to fix linking errors
